refactored getting page width
