+++ client/src/components/sections/DevSection.jsx
@@ -1,37 +1,19 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useQuery } from "@apollo/client";
 import { XMarkIcon } from "@heroicons/react/24/outline";
 
-import { GET_ARTS } from "../../utils/queries";
+import { GET_DEVS } from "../../utils/queries";
 
 import Carousel from "../Carousel";
 
-function DevSection() {
-  const { loading, data } = useQuery(GET_ARTS);
-  const artThings = data ? data.artThings : [];
+function DevSection(props) {
+  const { loading, data } = useQuery(GET_DEVS);
+  const devThings = data ? data.devThings : [];
 
   const CardComponent = Card;
   const [selectedProject, setSelectedProject] = useState();
-  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
-  const [cardsPerSlide, setCardsPerSlide] = useState();
-
-  useEffect(() => {
-    const handleResize = () => {
-      setWindowWidth(window.innerWidth);
-    };
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, []);
-
-  useEffect(() => {
-    if (windowWidth >= 1200) {
-      setCardsPerSlide("4");
-    } else if (1200 > windowWidth && windowWidth > 800) {
-      setCardsPerSlide("3");
-    } else {
-      setCardsPerSlide("2");
-    }
-  }, [windowWidth]);
+  const cardsPerSlide = props.cardsPerSlide
+  const windowWidth = props.windowWidth
 
   function MoreSection({ thing }) {
     function handleClose() {
@@ -89,7 +71,7 @@ function DevSection() {
           <h1>loading...</h1>
         ) : (
           <Carousel
-            items={artThings}
+            items={devThings}
             cardsPerSlide={cardsPerSlide}
             cardContainerWidth={(windowWidth - 30) / cardsPerSlide}
             CardComponent={CardComponent}
