+++ client/src/pages/HomePage.jsx
@@ -1,4 +1,4 @@
-import { useRef, useEffect } from "react";
+import { useRef, useEffect, useState } from "react";
 import { useParams, Link } from "react-router-dom";
 import { ArrowDownIcon } from "@heroicons/react/24/outline";
 
@@ -17,6 +17,10 @@ function HomePage() {
 
   const params = useParams();
 
+  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
+  const [cardsPerSlide, setCardsPerSlide] = useState();
+
+  // Scroll sections into view when clicked in navigation
   useEffect(() => {
     const scrollSpot = params.section;
     switch (scrollSpot) {
@@ -37,6 +41,26 @@ function HomePage() {
         break;
     }
   }, [params]);
+
+  // Get the window width when resized
+  useEffect(() => {
+    const handleResize = () => {
+      setWindowWidth(window.innerWidth);
+    };
+    window.addEventListener("resize", handleResize);
+    return () => window.removeEventListener("resize", handleResize);
+  }, []);
+
+  // set cards per slide based on window width
+  useEffect(() => {
+    if (windowWidth >= 1200) {
+      setCardsPerSlide("4");
+    } else if (1200 > windowWidth && windowWidth > 800) {
+      setCardsPerSlide("3");
+    } else {
+      setCardsPerSlide("2");
+    }
+  }, [windowWidth]);
 
   return (
     <>
@@ -83,19 +107,19 @@ function HomePage() {
       ></img>
 
       <div className="test" ref={devSection}>
-        <DevSection></DevSection>
+        <DevSection cardsPerSlide={cardsPerSlide} windowWidth={windowWidth}></DevSection>
       </div>
       <div className="test" ref={designSection} id="design-section">
-        <DesignSection></DesignSection>
+        <DesignSection cardsPerSlide={cardsPerSlide} windowWidth={windowWidth}></DesignSection>
       </div>
       <div ref={artSection} id="artSec">
-        <ArtSection></ArtSection>
+        <ArtSection cardsPerSlide={cardsPerSlide} windowWidth={windowWidth}></ArtSection>
       </div>
       <div className="test" ref={musicSection}>
-        <MusicSection></MusicSection>
+        <MusicSection cardsPerSlide={cardsPerSlide} windowWidth={windowWidth}></MusicSection>
       </div>
       <div className="test" ref={craftSection}>
-        <CraftSection></CraftSection>
+        <CraftSection cardsPerSlide={cardsPerSlide} windowWidth={windowWidth}></CraftSection>
       </div>
     </>
   );
